Cleaning up the core (modified version of a patch by Jim Studt): moving pin definitions to program space to save RAM, changing core function arguments (e.g. pinMode(), digitalWrite()) to uint8_t, restoring old SREG after delayMicroseconds() instead of always enabling interrupts, etc.

+++ targets/arduino/wiring_shift.c
@@ -24,7 +24,8 @@
 
 #include "wiring_private.h"
 
-void shiftOut(int dataPin, int clockPin, int bitOrder, byte val) {
+void shiftOut(uint8_t dataPin, uint8_t clockPin, uint8_t bitOrder, byte val)
+{
 	int i;
 
 	for (i = 0; i < 8; i++)  {
